+++ ntex-util/src/services/keepalive.rs
@@ -1,5 +1,4 @@
-use std::task::{Context, Poll};
-use std::{cell::Cell, convert::Infallible, fmt, marker, time::Duration, time::Instant};
+use std::{cell::Cell, convert::Infallible, fmt, future::poll_fn, marker, task, time};
 
 use ntex_service::{Service, ServiceCtx, ServiceFactory};
 
@@ -73,7 +72,7 @@ pub struct KeepAliveService<R, E, F> {
     f: F,
     dur: Millis,
     sleep: Sleep,
-    expire: Cell<Instant>,
+    expire: Cell<time::Instant>,
     _t: marker::PhantomData<(R, E)>,
 }
 
@@ -111,23 +110,24 @@ where
     type Response = R;
     type Error = E;
 
-    fn poll_ready(&self, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
-        match self.sleep.poll_elapsed(cx) {
-            Poll::Ready(_) => {
+    async fn ready(&self, _: ServiceCtx<'_, Self>) -> Result<(), Self::Error> {
+        poll_fn(|cx| match self.sleep.poll_elapsed(cx) {
+            task::Poll::Ready(_) => {
                 let now = now();
-                let expire = self.expire.get() + Duration::from(self.dur);
+                let expire = self.expire.get() + time::Duration::from(self.dur);
                 if expire <= now {
-                    Poll::Ready(Err((self.f)()))
+                    task::Poll::Ready(Err((self.f)()))
                 } else {
                     let expire = expire - now;
                     self.sleep
                         .reset(Millis(expire.as_millis().try_into().unwrap_or(u32::MAX)));
                     let _ = self.sleep.poll_elapsed(cx);
-                    Poll::Ready(Ok(()))
+                    task::Poll::Ready(Ok(()))
                 }
             }
-            Poll::Pending => Poll::Ready(Ok(())),
-        }
+            task::Poll::Pending => task::Poll::Ready(Ok(())),
+        })
+        .await
     }
 
     async fn call(&self, req: R, _: ServiceCtx<'_, Self>) -> Result<R, E> {
@@ -138,6 +138,8 @@ where
 
 #[cfg(test)]
 mod tests {
+    use std::task::Poll;
+
     use super::*;
     use crate::future::lazy;
 
@@ -150,7 +152,7 @@ mod tests {
         assert!(format!("{:?}", factory).contains("KeepAlive"));
         let _ = factory.clone();
 
-        let service = factory.pipeline(&()).await.unwrap();
+        let service = factory.pipeline(&()).await.unwrap().bind();
         assert!(format!("{:?}", service).contains("KeepAliveService"));
 
         assert_eq!(service.call(1usize).await, Ok(1usize));